# webapp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solo Leveling API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/dungeons/enter/{rank}": {
      "post": {
        "tags": [
          "dungeons"
        ],
        "summary": "Enter Dungeon",
        "operationId": "enter_dungeon_api_dungeons_enter__rank__post",
        "parameters": [
          {
            "name": "rank",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rank"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterDungeon"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dungeons/task/{index}": {
      "post": {
        "tags": [
          "dungeons"
        ],
        "summary": "Dungeon Task",
        "operationId": "dungeon_task_api_dungeons_task__index__post",
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DungeonTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/class/awakening/{class_key}": {
      "post": {
        "tags": [
          "class"
        ],
        "summary": "Choose Awakening",
        "operationId": "choose_awakening_api_class_awakening__class_key__post",
        "parameters": [
          {
            "name": "class_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Class Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AwakeningResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/class/jobchange/{class_key}": {
      "post": {
        "tags": [
          "class"
        ],
        "summary": "Choose Jobchange",
        "operationId": "choose_jobchange_api_class_jobchange__class_key__post",
        "parameters": [
          {
            "name": "class_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Class Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobChangeResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AwakeningResult": {
        "properties": {
          "class_stage1": {
            "type": "string",
            "title": "Class Stage1"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          }
        },
        "type": "object",
        "required": [
          "class_stage1",
          "class_name"
        ],
        "title": "AwakeningResult"
      },
...
      "DungeonTask": {
        "properties": {
          "dungeon_progress": {
            "type": "integer",
            "title": "Dungeon Progress"
          },
          "total_enemies": {
            "type": "integer",
            "title": "Total Enemies"
          },
          "boss_available": {
            "type": "boolean",
            "title": "Boss Available"
          }
        },
        "type": "object",
        "required": [
          "dungeon_progress",
          "total_enemies",
          "boss_available"
        ],
        "title": "DungeonTask"
      },
      "Enemy": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "task": {
            "type": "string",
            "title": "Task"
          }
        },
        "type": "object",
        "required": [
          "name",
          "task"
        ],
        "title": "Enemy"
      },
      "EnterDungeon": {
        "properties": {
          "rank": {
            "type": "string",
            "title": "Rank"
          },
          "enemies": {
            "items": {
              "$ref": "#/components/schemas/Enemy"
            },
            "type": "array",
            "title": "Enemies"
          },
          "time_limit": {
            "type": "integer",
            "title": "Time Limit"
          },
          "dungeon_end": {
            "type": "string",
            "title": "Dungeon End"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "enemies",
          "time_limit",
          "dungeon_end"
        ],
        "title": "EnterDungeon"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobChangeResult": {
        "properties": {
          "class_stage2": {
            "type": "string",
            "title": "Class Stage2"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          }
        },
        "type": "object",
        "required": [
          "class_stage2",
          "class_name"
        ],
        "title": "JobChangeResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}